fix compiler warnings
git-svn-id: svn://localhost:3390/trunk/guinea-pig@1346 489f5105-f64b-4bcc-8a87-8213ade39870

+++ src/gridCPP.cc
@@ -1891,7 +1891,7 @@ void GRID::move_photons2(BEAM& beam,int ibeam,int i_slice, RNDM& hasard)
 	    }
 	    else {
 	      nd=(int)(tmp/0.1+1.0);
-	      int ds=tmp/nd;
+	      float ds=tmp/nd;
 	      tmp=0.0;
 	      for (i=0;i<nd;++i) {
 		if (hasard.rndm()<ds) {
